% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataInt.R
\name{DataInt}
\alias{DataInt}
\title{A function that abstract the k columns (2 step interpolated data) in AHIdat_ap
AHIdat_ap is a rda file from the package of AHItool (unpublished) created by Wesley Burr}
\usage{
DataInt(dat, key = "k", WarmStart = 4, WarmEnd = 9)
}
\arguments{
\item{dat}{data.frame from AHIdat_ap list}

\item{key}{the first char in the names of columns that will be abstracted}

\item{WarmStart}{the starting month for warm season}

\item{WarmEnd}{the ending month for warm season}
}
\value{
datInt A data.frame that has dat's columns with a name starts with "k" and ends with "0"
}
\description{
A function that abstract the k columns (2 step interpolated data) in AHIdat_ap
AHIdat_ap is a rda file from the package of AHItool (unpublished) created by Wesley Burr
}
\examples{
dat <- data.frame(Month=rep(1:12,50),k.pollutant1.log1=runif(600),date=runif(600),k.pol.log0=runif(600))
dat.Int <- DataInt(dat)
}
