% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LagPlot.R
\name{LagPlot}
\alias{LagPlot}
\title{Plot the lag scatter plots and the regression lines for time series}
\usage{
LagPlot(ts_data, nlag = 2, Rsqur = TRUE)
}
\arguments{
\item{ts_data}{data of the time series as a vector}

\item{nlag}{number of lagging; 2 by default}

\item{Rsqur}{logical value; whether the R^2 values for regression are added on the plot; TRUE by default}
}
\description{
Plot the lag scatter plots and the regression lines for time series
}
\examples{
data("AHIdat_ap")
CDname<- names(AHIdat_ap)
dat_ap <- AHIdat_ap[1]
names(dat_ap) <- NULL
dat_ap <- data.frame(dat_ap[1])
ts_data <- dat_ap$kO3.24hm.lag0
LagPlot(ts_data,nlag=2,Rsqur = TRUE) 
}
