% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkplot.R
\name{sparkplot}
\alias{sparkplot}
\title{sparkplot}
\usage{
sparkplot(dat, level.trans = 0.5, ...)
}
\arguments{
\item{dat}{the data.frame user will pass into the function}

\item{level.trans}{level of transparency for the color of lines}

\item{...}{pass arguments to the embedded functions}
}
\description{
This function plot transparent lines from the a data.frame, the level of transparency was an parameter of the function
}
\examples{
num_data =20  # number of data for each line
          dat<-data.frame(col1=runif(num_data),col2=runif(num_data),
                          col3=runif(num_data),col4=runif(num_data),
                          col5=runif(num_data),col6=runif(num_data))
          sparkplot(dat,level.trans = 0.3,
                    xlab="xaxis", ylab="dataframe column values",
                    main ="Transparent Lines / Spark Plot")
}
