% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WCBLocks.R
\name{WCBlocks}
\alias{WCBlocks}
\title{Segregate pollutants and temperature time series by Warm/Cold seasons in 'AHIdat_ap.rda' and
'AHIdat_default.rda'. The return will be ready for a classification problem.}
\usage{
WCBlocks(
  listnum = 1,
  PolCol = c("k03.8hm", "k03.24hm", "kN02.24hm"),
  TempCol = "Temp"
)
}
\arguments{
\item{listnum}{The number in the list of data file 'AHIdat_ap.rda' and 'AHIdat_default.rda', each number represent a CD name. ranging from 1 to 24. Default value is 1, meaning the CD of Halifax.}

\item{PolCol}{The columns' names of pollutant measurements of interests, default value is c('k03.8hm','k03.24hm','kN02.24hm').}

\item{TempCol}{The column's name of temperature, default value is "Temp".}
}
\value{
X The input (of the classification problem) is an array with 3 dimension: i, j, k. i is the index for the i-th inputs/observation, j is the day index in each season, k is the  index for channels. The range of i is 1632 (34(years)*24(CDs)*2(seasons)). The range of j is 182 (days in one season). The range of k is 4 channels (k03.8hm,k03.24hm,kN02.24hm, and temp)

Y The output (of the classification problem) is an array with 1 dimension, i.e. the index for each response. i is the same as i in the input. The i represent the i-th datum in the target data, and its range is 1632 (34(years)*24(CDs)*2(seasons)). Each response has3 values. The first value is the type of season W/C, the second value is the year, the third value is the name of CD.
}
\description{
Segregate pollutants and temperature time series by Warm/Cold seasons in 'AHIdat_ap.rda' and
'AHIdat_default.rda'. The return will be ready for a classification problem.
}
